+++ CHANGELOG.md
@@ -1,11 +1,23 @@
-# Change Log
+# Changelog
+
+## [0.0.3] - 16th March 2023
+
+### Added
+
+- Infer (some) props for extracted component
+- Use TypeScript if original file also uses TypeScript
+- Add shortcut `Ctr` + `Alt` + `E`
+- Format extracted component upon creation
 
 ## [0.0.2] - 12th August 2019
+
 ### Added
+
 - Improved formatting
 - Imports the created component in current component
 
 ## [0.0.3] - 16th August 2019
-### Added
-- Simple directory support
 
+### Added
+
+- Simple directory support